Added photos to some computer clubs + made a image slider in InfoActivity + some design changes

+++ app/src/main/java/com/example/cybermap/ComputerClubData.kt
@@ -1,4 +1,4 @@
 package com.example.cybermap
 
 class ComputerClubData (var _id: Int, var name: String, var address: String, var phone: String, var site: String,
-                        var hours:String, var isAvailableOnlineBooking: Short, var coordinates: ArrayList<Double>, var images: List<Int>)
+                        var hours:String, var isAvailableOnlineBooking: Int, var coordinates: ArrayList<Double>, var images: List<Int>)
